refactor(tests): Split UserService tests into focused modules
- Split large UserService.test.ts (683 lines) into three focused files
- UserService.auth.test.ts: Authentication operations (14 tests)
- UserService.profile.test.ts: Profile management operations (10 tests)
- UserService.admin.test.ts: Admin operations and integration tests (10 tests)
- Each file under 300 lines, meeting project complexity standards
- All 34 tests pass, maintaining 100% UserService coverage
- Addresses Codacy complexity warning while preserving functionality

Resolves Codacy issue: file-nloc-medium
